+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_document_model_administration_client.py
@@ -28,7 +28,7 @@ from ._models import (
     ModelBuildMode,
     DocumentModelDetails,
     DocumentModelSummary,
-    DocumentModelOperationDetails,
+    OperationDetails,
     DocumentModelOperationSummary,
     ResourceDetails,
     TargetAuthorization,
@@ -457,7 +457,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBase):
         )
 
     @distributed_trace
-    def get_operation(self, operation_id: str, **kwargs: Any) -> DocumentModelOperationDetails:
+    def get_operation(self, operation_id: str, **kwargs: Any) -> OperationDetails:
         """Get a document model operation by its ID.
 
         Get a document model operation associated with the Form Recognizer resource.
@@ -465,8 +465,8 @@ class DocumentModelAdministrationClient(FormRecognizerClientBase):
         the document model can be accessed using the :func:`~get_model` or :func:`~list_models` APIs.
 
         :param str operation_id: The operation ID.
-        :return: DocumentModelOperationDetails
-        :rtype: ~azure.ai.formrecognizer.DocumentModelOperationDetails
+        :return: OperationDetails
+        :rtype: ~azure.ai.formrecognizer.OperationDetails
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. admonition:: Example:
@@ -482,7 +482,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBase):
         if not operation_id:
             raise ValueError("'operation_id' cannot be None or empty.")
 
-        return DocumentModelOperationDetails._from_generated(
+        return OperationDetails._from_generated(
             self._client.get_operation(operation_id, **kwargs),
             api_version=self._api_version,
         )