+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_repr.py
@@ -509,7 +509,7 @@ class TestRepr():
         assert repr(model) == model_repr
 
     def test_model_operation(self, document_analysis_error, document_model):
-        model = _models.DocumentModelOperationDetails(
+        model = _models.OperationDetails(
                 api_version="2022-08-31",
                 tags={"awesome": "tag"},
                 operation_id="id",
@@ -522,7 +522,7 @@ class TestRepr():
                 error=document_analysis_error[0],
                 result=document_model[0],
             )
-        model_repr = "DocumentModelOperationDetails(operation_id={}, status={}, percent_completed={}, created_on={}, last_updated_on={}, kind={}, resource_location={}, result={}, error={}, api_version={}, tags={})".format(
+        model_repr = "OperationDetails(operation_id={}, status={}, percent_completed={}, created_on={}, last_updated_on={}, kind={}, resource_location={}, result={}, error={}, api_version={}, tags={})".format(
                     "id",
                     "succeeded",
                     99,